+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-
+- Add `GetTimeOffs()` to handle `GET /company/time-offs`
+- Implement basic request handling and `accessToken` rotation
+- Add `Auth()` to handle `POST /auth`
+- Add Personio API v1 client
 
 [Unreleased]: https://github.com/giantswarm/personio-go/tree/master